rename: ApplicationException 를 별도 파일(base)로 분리
