+++ src/common/filters/http-exception/http-exception.filter.ts
@@ -1,6 +1,6 @@
 import { Catch, ExceptionFilter, ArgumentsHost } from '@nestjs/common';
 import { Response } from 'express';
-import { ApplicationException } from 'src/common/types/error/application-exceptions';
+import { ApplicationException } from 'src/common/types/error/application-exceptions.base';
 
 @Catch(ApplicationException)
 export class HttpExceptionFilter<T extends ApplicationException>
