+++ src/module/iam/guards/authentication.guard.ts
@@ -3,7 +3,7 @@ import { Reflector } from '@nestjs/core';
 import { AUTH_TYPE_KEY } from '../decorators/auth.decorator';
 import { AuthType } from '../enums/auth-type.enum';
 import { AccessTokenGuard } from './access-token.guard';
-import { ApplicationException } from 'src/common/types/error/application-exceptions';
+import { ApplicationException } from 'src/common/types/error/application-exceptions.base';
 
 @Injectable()
 export class AuthenticationGuard implements CanActivate {
